Added skia as zipped precompiled binaries (win64). Windows platform-specific code ready

+++ Win32Tests/TestClientArea/TestClientArea.cpp
@@ -441,5 +441,5 @@ int WINAPI WinMain(HINSTANCE hInstance, HINSTANCE hPrevInstance,
     TranslateMessage(&msg);
     DispatchMessage(&msg);
   }
-  return msg.wParam;
+  return static_cast<int>(msg.wParam);
 }
